tutorial_game/tutorial_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.24.3
gymnasium==0.29.1
minigrid==2.3.1 
python-dotenv
Pillow==11.0.0
sqlalchemy==2.0.23
Flask==3.0.3
Flask_SocketIO==5.3.6
flask_sqlalchemy==3.1.1
PyMySQL==1.1.1

